+++ src/Symfony/Component/Mailer/EventListener/SmimeEncryptedMessageListener.php
@@ -1,0 +1,47 @@
+<?php
+
+/*
+ * This file is part of the Symfony package.
+ *
+ * (c) Fabien Potencier <fabien@symfony.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Symfony\Component\Mailer\EventListener;
+
+use Symfony\Component\EventDispatcher\EventSubscriberInterface;
+use Symfony\Component\Mailer\Event\MessageEvent;
+use Symfony\Component\Mime\Crypto\SMimeEncrypter;
+use Symfony\Component\Mime\Message;
+
+/**
+ * Encrypts the message.
+ *
+ * @author Elías Fernández
+ */
+class SmimeEncryptedMessageListener implements EventSubscriberInterface
+{
+    public function __construct(
+        private SMimeEncrypter $encrypter,
+    ) {
+    }
+
+    public function onMessage(MessageEvent $event): void
+    {
+        $message = $event->getMessage();
+        if (!$message instanceof Message) {
+            return;
+        }
+
+        $event->setMessage($this->encrypter->encrypt($message));
+    }
+
+    public static function getSubscribedEvents(): array
+    {
+        return [
+            MessageEvent::class => ['onMessage', -128],
+        ];
+    }
+}